Moved more code around into separate packages, moved some piece and movement related things the piece package Added operation to fetch possible moves for a given piece
