+++ src/main/java/com/example/chess/game/GameService.java
@@ -4,11 +4,15 @@ import com.example.chess.game.exception.ChessException;
 import com.example.chess.game.exception.GameEndedException;
 import com.example.chess.game.exception.UnknownGameException;
 import com.example.chess.game.model.*;
+import com.example.chess.game.piece.MoveHelper;
+import com.example.chess.game.piece.Piece;
 import com.example.chess.game.util.BoardTemplate;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import java.util.Collection;
+import java.util.Collections;
+import java.util.stream.Collectors;
 
 @Component
 @AllArgsConstructor
@@ -28,7 +32,21 @@ public class GameService {
         return gameRepository.getGame(gameId).orElseThrow(() -> new UnknownGameException(gameId));
     }
 
-    public MovePieceResult movePiece(String gameId, String locationFromStr, String locationToStr)  throws ChessException {
+    public Collection<String> getPossibleMovesForLocation(String gameId, String locationStr) {
+        Game game = getGameState(gameId);
+        Location location = Location.fromText(locationStr);
+        Piece piece = game.getPieceAtLocation(location);
+        if (piece == null) {
+            return Collections.emptyList();
+        }
+
+        Collection<Location> possibleMoves = MoveHelper.getPossibleMoves(game, piece, location);
+        return possibleMoves.stream()
+                .map(Location::toString)
+                .collect(Collectors.toList());
+    }
+
+    public MovePieceResult movePiece(String gameId, String locationFromStr, String locationToStr) throws ChessException {
         Game game = getGameState(gameId);
 
         if (game.getState() == GameState.CHECKMATE) {
@@ -41,7 +59,7 @@ public class GameService {
         /// move piece (game, from, to)
         Collection<String> moveConsequences = game.movePiece(locationFrom, locationTo);
 
-        // verify game state, check/checkmate?
+        // update the game state
         moveConsequences.addAll(game.checkGameEndCondition());
 
         game.rotateCurrentPlayer();
@@ -52,5 +70,4 @@ public class GameService {
                 .consequences(moveConsequences)
                 .build();
     }
-
 }
