+++ src/main/java/com/example/chess/game/http/GameRestController.java
@@ -1,18 +1,20 @@
-package com.example.chess.game;
+package com.example.chess.game.http;
 
+import com.example.chess.game.GameService;
 import com.example.chess.game.exception.ChessException;
-import com.example.chess.game.model.Game;
-import com.example.chess.game.model.MovePieceRequest;
+import com.example.chess.game.Game;
 import com.example.chess.game.model.MovePieceResult;
-import com.example.chess.game.util.GameStateVisualizer;
+import com.example.chess.game.util.GameVisualizer;
 import lombok.AllArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Collection;
+
 @RestController
 @RequestMapping("/chess/game/")
 @AllArgsConstructor
-public class GameController {
+public class GameRestController {
 
     private GameService gameService;
 
@@ -42,8 +44,18 @@ public class GameController {
             produces = MediaType.TEXT_PLAIN_VALUE
     )
     @ResponseBody
-    public String visualizeGameState(@PathVariable String gameId) throws ChessException {
-        return GameStateVisualizer.visualizeGameBoard(gameService.getGameState(gameId));
+    public String visualizeGame(@PathVariable String gameId) throws ChessException {
+        return GameVisualizer.visualizeGame(gameService.getGameState(gameId));
+    }
+
+    @RequestMapping(
+            method = RequestMethod.GET,
+            path = "/{gameId}/moves",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    @ResponseBody
+    public Collection<String> getPieceMoves(@PathVariable String gameId, @RequestParam("location") String location) throws ChessException {
+        return gameService.getPossibleMovesForLocation(gameId, location);
     }
 
     @RequestMapping(
