+++ src/main/java/com/example/chess/game/http/MovePieceRequest.java
@@ -1,4 +1,4 @@
-package com.example.chess.game.model;
+package com.example.chess.game.http;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
